feat : 비슷한 ootd 조회(유사 스타일검색)시 ootdImage 가 아닌 ootd 를 반환하도록 수정
