+++ src/main/java/zip/ootd/ootdzip/ootd/data/OotdGetSimilarRes.java
@@ -2,7 +2,7 @@ package zip.ootd.ootdzip.ootd.data;
 
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import zip.ootd.ootdzip.ootdimage.domain.OotdImage;
+import zip.ootd.ootdzip.ootd.domain.Ootd;
 
 @Data
 @NoArgsConstructor
@@ -12,8 +12,8 @@ public class OotdGetSimilarRes {
 
     private String image;
 
-    public OotdGetSimilarRes(OotdImage ootdImage) {
-        this.id = ootdImage.getOotd().getId();
-        this.image = ootdImage.getImageUrl();
+    public OotdGetSimilarRes(Ootd ootd) {
+        this.id = ootd.getId();
+        this.image = ootd.getFirstImage();
     }
 }
